- cleaned up ERD pylance issues

+++ gehomesdk/erd/converters/ac/erd_wac_demand_response_power_converter.py
@@ -2,8 +2,8 @@ from ..abstract import ErdReadOnlyConverter
 from ..primitives import *
 from gehomesdk.erd.values.ac import *
 
-class ErdWacDemandResponsePowerConverter(ErdReadOnlyConverter[int]):
-    def erd_decode(self, value: str) -> int:
+class ErdWacDemandResponsePowerConverter(ErdReadOnlyConverter[float]):
+    def erd_decode(self, value: str) -> float:
         try:
             return float(erd_decode_int(value[:4])) / 1000.0
         except:
